web7

+++ src/controllers/users.controller.ts
@@ -1,5 +1,7 @@
 import User from "../models/User";
 import type { Response, Request } from "express";
+import jwt from "jsonwebtoken";
+import { JWT_SECRET } from "../config/constants";
 
 export const createUser = async (req: Request, res: Response) => {
     try {
@@ -21,14 +23,57 @@ export const getUsers = async (req: Request, res: Response) => {
     }
 }
 
-export const deleteUser = async (req: Request, res: Response) => {
-    const { userId } = req.body;
+export const updateUser = async (req: Request, res: Response): Promise<void> => {
     try {
-        const deleteUser = await User.findByIdAndDelete(userId);
-        if (!deleteUser) {
-            res.status(404).json({ message: "user not found" });
+
+        const { email } = req.body;
+        if (!email) {
+            res.status(400).json({ message: "Email is required for updating user" });
+            return;
         }
-        res.status(200).json(deleteUser);
+
+        const updateData = { ...req.body };
+        console.log(updateData);
+        delete updateData.email;
+
+        const updatedUser = await User.findOneAndUpdate(
+            { email },
+            updateData,
+            { new: true }
+        );
+
+        if (!updatedUser) {
+            res.status(404).json({ message: "User not found" });
+            return;
+        }
+
+        res.status(200).json(updatedUser);
+    } catch (e) {
+        res.status(400).json({ message: (e as Error).message });
+    }
+};
+
+export const deleteUser = async (req: Request, res: Response) => {
+    const authHeader = req.headers.authorization
+
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+        res.status(401).json({ message: "Authorization required" });
+        return;
+    }
+    const token = authHeader.split(" ")[1];
+    try {
+        const decoded = jwt.verify(token, JWT_SECRET);
+
+        console.log(decoded);
+    } catch (e) {
+        res.status(401).json({ message: "Invalid token" });
+        return;
+    }
+    try {
+        console.log(req.body);
+        const { userId } = req.body;
+        const result = await User.findByIdAndDelete(userId);
+        res.status(200).json(result);
     } catch (e) {
         res.status(400).json({ message: (e as Error).message })
     }
